+++ SmartDeviceLink/SDLSRGBColor.m
@@ -10,7 +10,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 @implementation SDLSRGBColor
 
-- (instancetype)initWithRed:(UInt32)red green:(UInt32)green blue:(UInt32)blue {
+- (instancetype)initWithRed:(UInt8)red green:(UInt8)green blue:(UInt8)blue {
     self = [self init];
     if(!self) {
         return nil;
